pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-rl"
version = "0.1.0"
description = "Agentic RLVR PoC: ART vs Agent Lightning on tau-bench"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0",
    "pydantic>=2.0",
    "pyyaml",
    "litellm",
    "wandb",
    "numpy",
    "pandas",
    "matplotlib",
]

[project.optional-dependencies]
tau = ["tau2 @ git+https://github.com/amazon-agi/tau2-bench-verified"]
art = ["openpipe-art"]
agl = ["agentlightning[verl]"]
dev = ["pytest", "pytest-asyncio", "ruff"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
